+++ src/components/Form.tsx
@@ -1,6 +1,13 @@
 import React, { FC, useState } from 'react';
 import styled from 'styled-components';
-import { TextInput } from 'react-native';
+import { TextInput, Button } from 'react-native';
+
+import { useDispatch } from 'react-redux';
+
+import { setNewElemToDoList } from '../actions/todolistActions';
+import { ISingleElementList } from '../entities/todoSingleEl';
+
+import Colors from '../constans/Colors';
 
 const Wrapper = styled.View`
     margin: 80px 20px 0 20px;
@@ -9,16 +16,40 @@ const Wrapper = styled.View`
 const CustomTextInput = styled.TextInput`
     border: 1px solid black;
     padding: 10px;
+    margin: 0 0 20px 0;
     color: black;
     width: 100%;
 `;
 
-const Form: FC = props => {
-    const [nameInput, setNameInput] = useState<string>(''); //25:53 - https://web.microsoftstream.com/video/20d79845-4c2b-45fb-8621-bea20563a062
+type setNewElemToDoList = ReturnType<typeof setNewElemToDoList>;
+
+const Form: FC<{switchView(formView: boolean)}> = props => {
+    const dispatch = useDispatch();
+    const [nameInput, setNameInput] = useState<string>('');
+    const [descInput, setDescInput] = useState<string>('');
+
+    const nameValueChange = (txt) => {
+        setNameInput(txt.nativeEvent.text);
+    }
+    
+    const descValueChange = (txt) => {
+        setDescInput(txt.nativeEvent.text);
+    }
+
+    const saveData = () => {
+        dispatch<setNewElemToDoList>(setNewElemToDoList({
+            name: nameInput,
+            description: descInput
+        } as ISingleElementList
+        ));
+        props.switchView(false)
+    }
+
     return (
         <Wrapper>
-            <CustomTextInput value={} onChange={} placeholder="Name"/>
-            <CustomTextInput value={} onChange={} placeholder="Description"/>
+            <CustomTextInput value={nameInput} onChange={nameValueChange} placeholder="Name"/>
+            <CustomTextInput value={descInput} onChange={descValueChange} placeholder="Description"/>
+            <Button title="Save" onPress={saveData}/>
         </Wrapper>
     )
 };