+++ src/constans/Colors.ts
@@ -1,5 +1,8 @@
 export default {
-    main: '#A9DDF2',
+    dark: '#4A87B2',
+    medium: '#6EB4E1',
+    light: '#8FCAEB',
+    inactive: '#878787',
     white: '#FFF',
     black: '#000'
 };