+++ CHANGELOG.md
@@ -4,12 +4,86 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2020-02-25
-
-### Changes
-
-    - 
+## [0.1] - 03-03-2020
 
 ### Added
 
     - Initial app config
+
+
+## [0.1.1] - 10-03-2020
+
+### Changes
+
+    - Changed splash screen icon
+
+
+## [0.2] - 17-03-2020
+
+### Added
+
+    - Bottom Navigation
+    - 3 screens for the navigation
+
+
+## [0.2.1] - 24-03-2020
+
+### Changes
+
+    - Bottom Navigation now has icons
+
+
+## [0.2.2] - 07-04-2020
+
+### Changes
+
+    - Bottom Navigation now properly takes you to correct screen
+
+
+## [0.3] - 14-04-2020
+
+### Changes
+
+    - Title screen now has an image and text
+
+
+## [0.3.1] - 21-04-2020
+
+### Changes
+
+    - Text screen now has an image and text
+
+
+## [0.3.2] - 28-04-2020
+
+### Changes
+
+    - Text screen is now a ScrollView that lets user scroll the text
+
+
+## [0.4] - 05-05-2020
+
+### Added
+
+    - Form that will be used to create tasks
+
+
+## [0.4.1] - 12-05-2020 19 26
+
+### Added
+
+    - Functionality to add tasks from code
+
+
+## [0.4.2] - 19-05-2020
+
+### Changes
+
+    - You can now add tasks from application itself
+
+
+## [0.5] - 26-05-2020
+
+### Changes
+
+    - Color overhaul